% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_install.r
\name{clean_install}
\alias{clean_install}
\title{Perform a clean install of geospatial package dependencies}
\usage{
clean_install(
  geo_pkgs = c("leaflet", "rgdal", "raster", "sp", "terra", "sf"),
  binary_repo = "https://ppm.publichealthscotland.org/all-r/__linux__/centos7/latest"
)
}
\arguments{
\item{geo_pkgs}{specify the geospatial packages for which dependencies will
be detected and removed}

\item{binary_repo}{url to specify a repo for package binaries}
}
\value{
invisible, function is called for its side-effects
}
\description{
Recursively detects geospatial package dependencies and removes
them before performing a fresh installation
}
