% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_env.r
\name{setup_env}
\alias{setup_env}
\title{Set LD_LIBRARY_PATH  and sets environment variables to facilitate use of
geospatial packages}
\usage{
setup_env()
}
\value{
invisible, function is called for its side-effects
}
\description{
Appends;
\itemize{
\item \verb{/usr/gdal<version>/lib}
\item \verb{/usr/proj<version>/lib}
}

to the users current LD_LIBRARY_PATH value.

Sets;
\itemize{
\item \verb{GDAL_DATA = /usr/gdal<version>/share/gdal}
\item \verb{PKG_CONFIG_PATH = /usr/proj<version>/lib/pkgconfig}
}

as environment variables.

Auto-detection of a working version of system dependencies is performed.
}
